# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Template
# during dev, should point to your local machine
host: localhost:10049
...
basePath: /template
...
schemes:
  # tip: remove http to make production-grade
  - http
  #- https
# format of bodies a client can send (Content-Type)
consumes:
  - multipart/form-data
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain
paths:
...
  /v1/batch:
    # binds a127 app logic to a route
    x-swagger-router-controller: controller
    get:
      description: Retrive a list of Batches
      operationId: v1_batchList
      parameters:
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of accounts per page 
        - name: filter
          in: query
          type: string
        - name: select
          in: query
          type: string
        - name : sort
          in : query
          type : string
      responses:
        200:
          description: lien lists
          schema:
            $ref: "#/definitions/TemplateDefinition1"
        400:
          description: Bad parameters
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: No transaction available with the given parameter set.
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /v1/batch/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: controller
    get:
      description: get batch by id
      operationId: v1_batchShow
      parameters:
        - name: id
          in: path
          type: string 
          required : true
          description: Page number of the request
        - name: filter
          in: query
          type: string
        - name: select
          in: query
          type: string
      responses:
        200:
          description: lien lists
          schema:
            $ref: "#/definitions/TemplateDefinition1"
        400:
          description: Bad parameters
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: No transaction available with the given parameter set.
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /v1/batch/process:
    # binds a127 app logic to a route
    x-swagger-router-controller: controller
    post:
      description: process the Pending Batches
      operationId: v1_batchProcess
      responses:
        200:
          description: batch procssed 
          schema:
            $ref: "#/definitions/TemplateDefinition1"
        400:
          description: Bad parameters
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: No transaction available with the given parameter set.
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    properties:
      error:
        type: string
  TemplateDefinition1:
    properties:
      _id:
        type: string